refactor(user): finished
