--- src-user/DependencyInjection/FOSUserExtension.php
@@ -1,109 +1,0 @@
-<?php
-
-/*
- * This file is part of the FOSUserBundle package.
- *
- * (c) FriendsOfSymfony <http://friendsofsymfony.github.com/>
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- */
-
-namespace FOS\UserBundle\DependencyInjection;
-
-use Symfony\Component\Config\Definition\Processor;
-use Symfony\Component\Config\FileLocator;
-use Symfony\Component\DependencyInjection\Alias;
-use Symfony\Component\DependencyInjection\ContainerBuilder;
-use Symfony\Component\DependencyInjection\Loader\XmlFileLoader;
-use Symfony\Component\HttpKernel\DependencyInjection\Extension;
-
-class FOSUserExtension extends Extension
-{
-    /**
-     * @var array
-     */
-    private static $doctrineDrivers = [
-        'orm' => [
-            'registry' => 'doctrine',
-            'tag' => 'doctrine.event_subscriber',
-        ],
-    ];
-
-    private $mailerNeeded = false;
-    private $sessionNeeded = false;
-
-    /**
-     * {@inheritdoc}
-     */
-    public function load(array $configs, ContainerBuilder $container)
-    {
-        $processor = new Processor();
-        $configuration = new Configuration();
-
-        $config = $processor->processConfiguration($configuration, $configs);
-
-        $loader = new XmlFileLoader($container, new FileLocator(__DIR__.'/../Resources/config'));
-
-        if ('custom' !== $config['db_driver']) {
-            if (isset(self::$doctrineDrivers[$config['db_driver']])) {
-                $container->setAlias('fos_user.doctrine_registry', new Alias(self::$doctrineDrivers[$config['db_driver']]['registry'], false));
-            } else {
-                $loader->load(\sprintf('%s.xml', $config['db_driver']));
-            }
-            $container->setParameter($this->getAlias().'.backend_type_'.$config['db_driver'], true);
-        }
-
-        $this->remapParametersNamespaces($config, $container, [
-            '' => [
-                'db_driver' => 'fos_user.storage',
-                'firewall_name' => 'fos_user.firewall_name',
-                'model_manager_name' => 'fos_user.model_manager_name',
-                'user_class' => 'fos_user.model.user.class',
-            ],
-        ]);
-
-        if ($this->sessionNeeded) {
-            // Use a private alias rather than a parameter, to avoid leaking it at runtime (the private alias will be removed)
-            $container->setAlias('fos_user.session', new Alias('session', false));
-        }
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getNamespace()
-    {
-        return 'http://friendsofsymfony.github.io/schema/dic/user';
-    }
-
-    protected function remapParameters(array $config, ContainerBuilder $container, array $map)
-    {
-        foreach ($map as $name => $paramName) {
-            if (\array_key_exists($name, $config)) {
-                $container->setParameter($paramName, $config[$name]);
-            }
-        }
-    }
-
-    protected function remapParametersNamespaces(array $config, ContainerBuilder $container, array $namespaces)
-    {
-        foreach ($namespaces as $ns => $map) {
-            if ($ns) {
-                if (!\array_key_exists($ns, $config)) {
-                    continue;
-                }
-                $namespaceConfig = $config[$ns];
-            } else {
-                $namespaceConfig = $config;
-            }
-            if (\is_array($map)) {
-                $this->remapParameters($namespaceConfig, $container, $map);
-            } else {
-                foreach ($namespaceConfig as $name => $value) {
-                    $container->setParameter(\sprintf($map, $name), $value);
-                }
-            }
-        }
-    }
-}
