--- src-user/Resources/config/doctrine-mapping/User.orm.xml
@@ -1,33 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<doctrine-mapping xmlns="http://doctrine-project.org/schemas/orm/doctrine-mapping"
-                  xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-                  xsi:schemaLocation="http://doctrine-project.org/schemas/orm/doctrine-mapping
-                  http://doctrine-project.org/schemas/orm/doctrine-mapping.xsd">
-
-    <mapped-superclass name="FOS\UserBundle\Model\User">
-
-        <field name="username" column="username" type="string" length="180" />
-
-        <field name="usernameCanonical" column="username_canonical" type="string" length="180" unique="true" />
-
-        <field name="email" column="email" type="string" length="180" />
-
-        <field name="emailCanonical" column="email_canonical" type="string" length="180" unique="true" />
-
-        <field name="enabled" column="enabled" type="boolean" />
-
-        <field name="salt" column="salt" type="string" nullable="true" />
-
-        <field name="password" column="password" type="string" />
-
-        <field name="lastLogin" column="last_login" type="datetime" nullable="true" />
-
-        <field name="confirmationToken" column="confirmation_token" type="string" length="180" unique="true" nullable="true" />
-
-        <field name="passwordRequestedAt" column="password_requested_at" type="datetime" nullable="true" />
-
-        <field name="roles" column="roles" type="array" />
-
-    </mapped-superclass>
-
-</doctrine-mapping>
