--- src-user/Resources/config/storage-validation/orm.xml
@@ -1,27 +1,0 @@
-<?xml version="1.0" ?>
-<constraint-mapping xmlns="http://symfony.com/schema/dic/constraint-mapping"
-    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-    xsi:schemaLocation="http://symfony.com/schema/dic/constraint-mapping
-        http://symfony.com/schema/dic/constraint-mapping/constraint-mapping-1.0.xsd">
-
-    <class name="FOS\UserBundle\Model\User">
-        <constraint name="Symfony\Bridge\Doctrine\Validator\Constraints\UniqueEntity">
-            <option name="fields">usernameCanonical</option>
-            <option name="errorPath">username</option>
-            <option name="message">fos_user.username.already_used</option>
-            <option name="groups">
-                <value>Profile</value>
-            </option>
-        </constraint>
-
-        <constraint name="Symfony\Bridge\Doctrine\Validator\Constraints\UniqueEntity">
-            <option name="fields">emailCanonical</option>
-            <option name="errorPath">email</option>
-            <option name="message">fos_user.email.already_used</option>
-            <option name="groups">
-                <value>Profile</value>
-            </option>
-        </constraint>
-    </class>
-
-</constraint-mapping>
