+++ src/Entity/User.php
@@ -5,22 +5,23 @@ namespace EMS\CoreBundle\Entity;
 use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\Common\Collections\Collection;
 use Doctrine\ORM\Mapping as ORM;
+use EMS\CoreBundle\Roles;
 use EMS\CoreBundle\Security\CoreLdapUser;
-use FOS\UserBundle\Model\User as BaseUser;
+use EMS\Helpers\Standard\Type;
 
 /**
  * @ORM\Entity
  * @ORM\Table(name="`user`")
  * @ORM\HasLifecycleCallbacks()
  */
-class User extends BaseUser implements UserInterface, EntityInterface
+class User implements UserInterface, EntityInterface
 {
     /**
      * @ORM\Id
      * @ORM\Column(type="integer")
      * @ORM\GeneratedValue(strategy="AUTO")
      */
-    protected $id;
+    private ?int $id = null;
 
     /**
      * @var \DateTime
@@ -107,7 +108,7 @@ class User extends BaseUser implements UserInterface, EntityInterface
     private $authTokens;
 
     /**
-     * @ORM\Column(name="locale", type="string", nullable=false)
+     * @ORM\Column(name="locale", type="string", nullable=false, options={"default":"en"})
      */
     private string $locale;
 
@@ -116,16 +117,74 @@ class User extends BaseUser implements UserInterface, EntityInterface
      */
     private ?string $localePreferred = null;
 
+    /**
+     * @ORM\Column(name="username", type="string", length=180)
+     */
+    private string $username = '';
+
+    /**
+     * @ORM\Column(name="username_canonical", type="string", length=180, unique=true)
+     */
+    private ?string $usernameCanonical = null;
+
+    /**
+     * @ORM\Column(name="email", type="string", length=180)
+     */
+    private string $email = '';
+
+    /**
+     * @ORM\Column(name="email_canonical", type="string", length=180, unique=true)
+     */
+    private ?string $emailCanonical = null;
+
+    /**
+     * @ORM\Column(name="enabled", type="boolean")
+     */
+    private bool $enabled;
+
+    /**
+     * @ORM\Column(name="salt", type="string", nullable=true)
+     */
+    private ?string $salt = null;
+
+    /**
+     * @ORM\Column(name="password", type="string")
+     */
+    private ?string $password = null;
+    private ?string $plainPassword = null;
+
+    /**
+     * @ORM\Column(name="last_login", type="datetime", nullable=true)
+     */
+    private ?\DateTime $lastLogin = null;
+
+    /**
+     * @ORM\Column(name="confirmation_token", type="string", length=180, unique=true, nullable=true)
+     */
+    private ?string $confirmationToken = null;
+
+    /**
+     * @ORM\Column(name="password_requested_at", type="datetime", nullable=true)
+     */
+    private ?\DateTime $passwordRequestedAt = null;
+
+    /**
+     * @var string[]
+     *
+     * @ORM\Column(name="roles", type="array")
+     */
+    private array $roles;
+
     private const DEFAULT_LOCALE = 'en';
 
     public function __construct()
     {
-        parent::__construct();
-
         $this->layoutBoxed = false;
         $this->sidebarCollapse = false;
         $this->sidebarMini = true;
         $this->authTokens = new ArrayCollection();
+        $this->enabled = false;
+        $this->roles = [];
         $this->locale = self::DEFAULT_LOCALE;
     }
 
@@ -512,5 +571,231 @@ class User extends BaseUser implements UserInterface, EntityInterface
         }
 
         return ($this->passwordRequestedAt->getTimestamp() + $ttl) > \time();
+    }
+
+    public function __toString(): string
+    {
+        return $this->getUsername();
+    }
+
+    public function addRole(string $role): void
+    {
+        $role = \strtoupper($role);
+        if (Roles::ROLE_USER === $role) {
+            return;
+        }
+
+        if (!\in_array($role, $this->roles, true)) {
+            $this->roles[] = $role;
+        }
+    }
+
+    public function serialize(): string
+    {
+        return \serialize([
+            $this->password,
+            $this->salt,
+            $this->usernameCanonical,
+            $this->username,
+            $this->enabled,
+            $this->id,
+            $this->email,
+            $this->emailCanonical,
+        ]);
+    }
+
+    public function unserialize(string $serialized): void
+    {
+        $data = \unserialize($serialized);
+
+        if (13 === \count($data)) {
+            // Unserializing a User object from 1.3.x
+            unset($data[4], $data[5], $data[6], $data[9], $data[10]);
+            $data = \array_values($data);
+        } elseif (11 === \count($data)) {
+            // Unserializing a User from a dev version somewhere between 2.0-alpha3 and 2.0-beta1
+            unset($data[4], $data[7], $data[8]);
+            $data = \array_values($data);
+        }
+
+        list(
+            $this->password,
+            $this->salt,
+            $this->usernameCanonical,
+            $this->username,
+            $this->enabled,
+            $this->id,
+            $this->email,
+            $this->emailCanonical
+            ) = $data;
+    }
+
+    public function eraseCredentials(): void
+    {
+        $this->plainPassword = null;
+    }
+
+    public function getId(): int
+    {
+        return Type::integer($this->id);
+    }
+
+    public function getUsername(): string
+    {
+        return $this->username;
+    }
+
+    public function getUsernameCanonical(): ?string
+    {
+        return $this->usernameCanonical;
+    }
+
+    public function getSalt(): ?string
+    {
+        return $this->salt;
+    }
+
+    public function getEmail(): string
+    {
+        return $this->email;
+    }
+
+    public function getEmailCanonical(): ?string
+    {
+        return $this->emailCanonical;
+    }
+
+    public function getPassword(): ?string
+    {
+        return $this->password;
+    }
+
+    public function getPlainPassword(): ?string
+    {
+        return $this->plainPassword;
+    }
+
+    public function getLastLogin(): ?\DateTime
+    {
+        return $this->lastLogin;
+    }
+
+    public function getConfirmationToken(): ?string
+    {
+        return $this->confirmationToken;
+    }
+
+    /**
+     * @return string[]
+     */
+    public function getRoles(): array
+    {
+        $roles = $this->roles;
+
+        // we need to make sure to have at least one role
+        $roles[] = Roles::ROLE_USER;
+
+        return \array_unique($roles);
+    }
+
+    public function hasRole(string $role): bool
+    {
+        return \in_array(\strtoupper($role), $this->getRoles(), true);
+    }
+
+    public function isSuperAdmin(): bool
+    {
+        return $this->hasRole(Roles::ROLE_SUPER_ADMIN);
+    }
+
+    public function removeRole(string $role): void
+    {
+        if (false !== $key = \array_search(\strtoupper($role), $this->roles, true)) {
+            unset($this->roles[$key]);
+            $this->roles = \array_values($this->roles);
+        }
+    }
+
+    public function setUsername(string $username): void
+    {
+        $this->username = $username;
+    }
+
+    public function setUsernameCanonical(?string $usernameCanonical): void
+    {
+        $this->usernameCanonical = $usernameCanonical;
+    }
+
+    public function setSalt(string $salt): void
+    {
+        $this->salt = $salt;
+    }
+
+    public function setEmail(string $email): void
+    {
+        $this->email = $email;
+    }
+
+    public function setEmailCanonical(?string $emailCanonical): void
+    {
+        $this->emailCanonical = $emailCanonical;
+    }
+
+    public function setEnabled(bool $enabled): void
+    {
+        $this->enabled = $enabled;
+    }
+
+    public function setPassword(?string $password): void
+    {
+        $this->password = $password;
+    }
+
+    public function setSuperAdmin(bool $superAdmin): void
+    {
+        if ($superAdmin) {
+            $this->addRole(Roles::ROLE_SUPER_ADMIN);
+        } else {
+            $this->removeRole(Roles::ROLE_SUPER_ADMIN);
+        }
+    }
+
+    public function setPlainPassword(?string $password): void
+    {
+        $this->plainPassword = $password;
+    }
+
+    public function setLastLogin(\DateTime $time = null): void
+    {
+        $this->lastLogin = $time;
+    }
+
+    public function setConfirmationToken(?string $confirmationToken): void
+    {
+        $this->confirmationToken = $confirmationToken;
+    }
+
+    public function setPasswordRequestedAt(\DateTime $date = null): void
+    {
+        $this->passwordRequestedAt = $date;
+    }
+
+    public function getPasswordRequestedAt(): ?\DateTime
+    {
+        return $this->passwordRequestedAt;
+    }
+
+    /**
+     * @param string[] $roles
+     */
+    public function setRoles(array $roles): self
+    {
+        $this->roles = [];
+
+        foreach ($roles as $role) {
+            $this->addRole($role);
+        }
+
+        return $this;
     }
 }
